requirements.txt
tensorflow-gpu
opencv-python~=4.5.3.56
icecream~=2.1.1
numpy~=1.19.5
click~=7.1.2
matplotlib~=3.3.4

tqdm~=4.62.0
mayavi~=4.7.3
plyfile~=0.7.4
scikit-learn~=0.24.2
numpy-quaternion
tensorflow~=2.5.0
svgpathtools